# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ACCweb api documentation",
        "title": "ACCWeb API documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ACCWeb project",
            "url": "https://github.com/assetto-corsa-web/accweb"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/assetto-corsa-web/accweb/blob/master/LICENSE"
        },
        "version": "1.19"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/metadata": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Metadata informations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "servers"
                ],
                "summary": "Returns server OS informations",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/servers": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "List all ACC dedicated servers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "servers"
                ],
                "summary": "List all ACC dedicated servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/app.ListServerItem"
                            }
                        }
                    }
                }
            }
        },
        "/servers/stop-all": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Stops all running ACC dedicated servers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "servers"
                ],
                "summary": "Stops all running ACC dedicated servers",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "app.ListServerItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "isRunning": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nrClients": {
                    "type": "integer"
                },
                "pid": {
                    "type": "integer"
                },
                "serverState": {
                    "$ref": "#/definitions/instance.ServerState"
                },
                "sessionPhase": {
                    "type": "string"
                },
                "sessionRemaining": {
                    "type": "integer"
                },
                "sessionType": {
                    "type": "string"
                },
                "tcpPort": {
                    "type": "integer"
                },
                "track": {
                    "type": "string"
                },
                "udpPort": {
                    "type": "integer"
                }
            }
        },
...
        "instance.ServerState": {
            "type": "string",
            "enum": [
                "offline",
                "starting",
                "stoping",
                "not_registered",
                "online"
            ],
            "x-enum-varnames": [
                "ServerStateOffline",
                "ServerStateStarting",
                "ServerStateStoping",
                "ServerStateNotRegistered",
                "ServerStateOnline"
            ]
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}